requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
python-multipart==0.0.20
python-magic==0.4.27
pillow==12.0.0
pydantic==2.12.4
numpy==2.2.6
opencv-python-headless==4.12.0.88
torch==2.9.1
torchvision==0.24.1
torchxrayvision==1.4.0
matplotlib==3.10.7

pytest==8.3.3
mypy==1.18.2
